remove int cast from spots

+++ bigstream/align.py
@@ -253,9 +253,9 @@ def feature_point_ransac_affine_align(
     # sort
     print('sorting spots', flush=True)
     sort_idx = np.argsort(fix_spots[:, 3])[::-1]
-    fix_spots = fix_spots[sort_idx, :3][:nspots].astype(int)
+    fix_spots = fix_spots[sort_idx, :3][:nspots]
     sort_idx = np.argsort(mov_spots[:, 3])[::-1]
-    mov_spots = mov_spots[sort_idx, :3][:nspots].astype(int)
+    mov_spots = mov_spots[sort_idx, :3][:nspots]
 
     # get contexts
     print('extracting contexts', flush=True)
